+++ templates/sql-ec2/tests-sql.yaml
@@ -8,7 +8,7 @@
   loadgenerator_hosts: &loadgenerator_hosts loadgenerators
   node_hosts: &node_hosts nodes
   driver: &driver hazelcast5
-  version: &version maven=5.3.0-SNAPSHOT
+  version: &version maven=5.3.1
   client_args: &client_args >
     -Xms10G
     -Xmx10G
